# public/postman/schemas/api.yml
openapi: 3.0.3
info:
  title: Arcane City
  version: 2025.06.09
  description: REST API for working with Arcane City data
  contact:
    name: Geoff Maddock
    email: geoff.maddock@gmail.com
servers:
  - url: https://arcane.city
    description: Production
  - url: https://stage.arcane.city
    description: Stage
  - url: https://dev.arcane.city
    description: Development
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    Forum:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          maxLength: 255
          description: A name of a forum
          example: Forum
        slug:
          type: string
          maxLength: 255
          description: A unique identifier name for the forum in kebab-case
          example: forum-slug
        short:
          type: string
          example: The short description of a forum
          description: A brief description of the forum
          maxLength: 255
        description:
          type: string
          example: An extensive breakdown of the forum
          description: An extensive breakdown of the forum
        visibility_id:
          type: integer
          readOnly: true
          example: 1
          description: Relation to the visibility table that defines the visibility of the blog
        created_at:
          type: string
          example: "2018-03-20T09:12:28Z"
          format: date-time
          description: Date and time that the entity type was created
          readOnly: true
        updated_at:
          type: string
          example: "2018-03-20T09:12:28Z"
          format: date-time
          description: Date and time that the entity type was last updated
          readOnly: true
    Forums:
      allOf: [$ref: "#/components/schemas/Pagination"]
      type: object
      properties:
        data:
          type: array
          description: List of the current page of threads
          items:
            "$ref": "#/components/schemas/Forum"
...
    Pagination:
      type: object
      properties:
        current_page:
          description: The current page in the paginated list
          type: integer
        first_page_url:
          description: URL for the first page
          type: string
          example: "https://dev.arcane.city/api/resources?page=1"
        from:
          description: First element in the current page
          type: integer
          example: 1
        last_page:
          description: Last page of results
          type: integer
          example: 1
        last_page_url:
          description: URL for the first page
          type: string
          example: "https://dev.arcane.city/api/resources?page=100"
        links:
          description: Array of links used to paginate the list of entities
          type: array
          items: {}
        next_page_url:
          description: URL for the next page
          type: string
          example: "https://dev.arcane.city/api/resources?page=2"
        path:
          description: Path for the primary route
          type: string
          example: "https://dev.arcane.city/api/resources"
        per_page:
          description: Number of results per page
          type: integer
          example: 25
        prev_page_url:
          description: URL for the previous page
          type: string
          example: "https://dev.arcane.city/api/resources?page=0"
        to:
          description: Last element in the current page
          type: integer
          example: 25
        total:
          description: Total number of results from the request
          type: integer
          example: 100
        data:
          description: List of the current page of the paginated entity
          type: array
          items: {}
security:
  - basicAuth: []
paths:
...
  /api/forums:
    get:
      tags:
        - forums
      summary: Get Forums
      operationId: getForums
      parameters:
        - name: filters[name]
          in: query
          required: false
          description: A filter query of the forum name
          schema:
            type: string
            example: Blog Title
        - name: limit
          in: query
          required: false
          description: The limit on the number of results to return per page
          schema:
            type: integer
            example: 25
        - name: page
          in: query
          required: false
          description: The page number to return
          schema:
            type: integer
            example: 1
        - name: sort
          in: query
          required: false
          description: A column to be used to sort the query
          schema:
            type: string
            example: name
        - name: direction
          in: query
          required: false
          description: A string indicating the sort direction of the query (asc or desc)
          schema:
            type: string
            example: asc
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forums"
    post:
      tags:
        - forums
      summary: Create forum
      operationId: createForum
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Forum"
      responses:
        "201":
          description: Successful response
          content:
            application/json: {}
  /api/forums/{forumId}:
    get:
      parameters:
        - name: forumId
          description: The unique identifier of the forum
          in: path
          required: true
          schema:
            description: The unique identifier of a forum
            type: integer
            readOnly: true
            example: 1
      tags:
        - forums
      summary: Get Forum
      operationId: getForumId
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forum"
    put:
      parameters:
        - name: forumId
          description: The unique identifier of the forum
          in: path
          required: true
          schema:
            description: The unique identifier of a forum
            type: integer
            readOnly: true
            example: 1
      tags:
        - forums
      summary: Update Forum (PUT — full replace)
      description: |
        Full replacement of the resource per REST PUT semantics. Optional scalar fields omitted from the body are reset to null, and any *_list relation arrays sync to exactly the supplied values — omitting a list detaches all rows for that relation. Use PATCH for partial updates.
      operationId: updateForumById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Forum"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forum"
    patch:
      parameters:
        - name: forumId
          description: The unique identifier of the forum
          in: path
          required: true
          schema:
            description: The unique identifier of a forum
            type: integer
            readOnly: true
            example: 1
      tags:
        - forums
      summary: Patch Forum (PATCH — partial update)
      description: |
        Partial update per REST PATCH semantics. Only the fields present in the body are modified; scalars and relations not included are left untouched. Validation rules from the corresponding *Request schema are applied as `sometimes` rules — if a field is supplied it must satisfy its constraints, but missing fields are not rejected. To clear a relation explicitly, send an empty array (e.g. `"tag_list": []`).
        Full replacement of the resource per REST PUT semantics. Optional scalar fields omitted from the body are reset to null, and any *_list relation arrays sync to exactly the supplied values — omitting a list detaches all rows for that relation. Use PATCH for partial updates.
      operationId: patchForumById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Forum"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forum"
    delete:
      parameters:
        - name: forumId
          description: The unique identifier of the forum
          in: path
          required: true
          schema:
            description: The unique identifier of a forum
            type: integer
            readOnly: true
            example: 1
      tags:
        - forums
      summary: Delete Forum
      operationId: deleteForumById
      responses:
        "204":
          description: Successful response
          content:
            application/json: {}
...